Move RHVoice_init_params, RHVoice_callback_types, RHVoice_callbacks to RHVoice.py
